# OpenAPI/coffeeshops.yaml
openapi: 3.1.1
info:
  title: CoffeeShops
  description: REST API для системы управления кофейнями, напитками и пользователями
  version: 1.0.0
  contact: {}
servers:
  - url: /api/v1
    description: Основной сервер API
    
#-----------------------------------------SCHEMAS-----------------------------------------------
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateResponse:
      type: object
      properties:
        id:
          type: string
          example: "9b165097-1c9f-4ea3-bef0-e505baa4ff63"
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Описание ошибки"
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        total:  #общее кол-во элем-ов во всех стр-цах
          type: integer
          example: 100
        page:   #тек. стр-ца
          type: integer
          example: 1
        limit:   #кол-во элементов на одной странице
          type: integer
          example: 20

    #-------CATEGORY----------------
...
    CoffeeShop:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор кофейни
        companyId:
          type: string
          format: uuid
          description: Идентификатор компании
        address:
          type: string
          description: Адрес кофейни
        workingHours:
          type: string
          description: Часы работы кофейни
      example:
        id: "123e4567-e89b-12d3-a456-426614174000"
        companyId: "123e4567-e89b-12d3-a456-426614174001"
        address: "Москва, Тверская, 10"
        workingHours: "08:00-22:00"
...
    Company:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор компании
          readOnly: true
        name:
          type: string
          description: Название компании кофеен
          maxLength: 100
        website:
          type: string
          description: URL веб-сайта компании
      example:
        id: "123e4567-e89b-12d3-a456-426614174000"
        name: "Starbucks"
        website: "https://www.starbucks.ru"
    CompanyData:
      type: object
      required:
        - name
        - website
      properties:
        name:
          type: string
          description: Название компании кофеен
          maxLength: 100
        website:
          type: string
          description: URL веб-сайта компании
      example:
        name: "Новая кофейная компания"
        website: "https://www.newcoffee.ru"


    #------------------------MENU ITEMS-------------------------
    Menu:
      type: object
      properties:
        companyId:
          type: string
          format: uuid
          description: Идентификатор компании
        drinkId:
          type: string
          format: uuid
          description: Идентификатор напитка
        price:
          type: number
          format: decimal
          description: Цена напитка
        size:
          type: number
          description: Объем в миллилитрах
      example:
        companyId: "123e4567-e89b-12d3-a456-426614174001"
        drinkId: "123e4567-e89b-12d3-a456-426614174002"
        price: 250.00
        volume: 350
...
  parameters:
    PageParam:
      name: page
      in: query
      description: Номер страницы для пагинации
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: LimitParam
      in: query
      description: Количество элементов на странице
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    DrinkId:
      name: drinkId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    CompanyId:
      name: companyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    OnlyFavoritesFilterQuery:
      name: onlyFavorites
      in: query
      required: false
      schema:
        type: boolean
      description: Фильтр только избранных элементов для текущего пользователя
  responses:
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Произошла внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Конфликт данных
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /companies:
    get:
      tags:
        - Companies
      summary: Получить все сети кофеен
      description: Возвращает пагинированный список кофеен
      operationId: GetAllCompanies
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Список сетей кофеен
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Company'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - Companies
      summary: Создать новую сеть кофеен
      description: Создает новую еть кофеен (только для модераторов и администраторов)
      operationId: AddCompany
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyData'
      responses:
        '200':
          description: Сеть кофеен успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # /companies/{drinkId}:
  #   get:
  #     tags:
  #       - Companies
  #     summary: Получить все сети кофеен, в меню которых есть определенный напиток 
  #     description: Возвращает пагинированный список кофеен, в меню которых есть определенный напиток 
  #     operationId: GetCompaniesByDrinkid
  #     parameters:
  #       - $ref: '#/components/parameters/DrinkId'
  #       - $ref: '#/components/parameters/PageParam'
  #       - $ref: '#/components/parameters/LimitParam'
  #     responses:
  #       '200':
  #         description: Список сетей кофеен, в меню которых есть определенный напиток 
  #         content:
  #           application/json:
  #             schema:
  #               allOf:
  #                 - $ref: '#/components/schemas/PaginatedResponse'
  #                 - type: object
  #                   properties:
  #                     data:
  #                       type: array
  #                       items:
  #                         $ref: '#/components/schemas/Company'
  #       '404':
  #         $ref: '#/components/responses/NotFound'
  /companies/{companyId}:
    delete:
      tags:
        - Companies
      summary: Удалить сеть кофеен
      description: Удаляет сеть кофеен по идентификатору (только для модераторов и администраторов)
      operationId: DeleteCompany
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CompanyId'
      responses:
        '200':
          description: Сеть кофеен успешно удалена
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
        - Companies
      summary: Получить сеть кофеен по ID
      description: Возвращает информацию о сети кофеен по ее идентификатору
      operationId: GetCompanyById
      parameters:
        - $ref: '#/components/parameters/CompanyId'
      responses:
        '200':
          description: Информация о сети кофеен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '404':
          $ref: '#/components/responses/NotFound'

        '500':
          $ref: '#/components/responses/InternalServerError'
  /companies/{companyId}/menu:
    get:
      tags:
        - Companies
      summary: Получить все позиции меню сети кофеен
      description: Получает список всех позиций меню сети кофеен
      operationId: GetAllMenuItems
      parameters:
        - $ref: '#/components/parameters/CompanyId'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Список позиций меню сети кофеен
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Menu'
        '404':
          $ref: '#/components/responses/NotFound'
    
    post:
      tags:
        - Companies
      summary: Добавить позицию в меню
      description: Добавляет новую позицию в меню конкретной компании
      operationId: AddMenuItem
      security:
        - bearerAuth: []  # Добавил аутентификацию для консистентности
      parameters:
        - name: companyId
          in: path
          required: true
          description: ID компании
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Menu'
      responses:
        '201':  # Изменил на 201 Created - более семантично для POST
          description: Позиция успешно добавлена в меню
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /companies/{companyId}/coffeeshops:
    get:
      tags:
        - Companies
      summary: Получить все кофейни сети кофеен
      description: Возвращает пагинированный список кофеен конкретной сети с фильтрацией
      operationId: GetCoffeeShopsByCompanyId
      parameters:
        - $ref: '#/components/parameters/CompanyId'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OnlyFavoritesFilterQuery'
      responses:
        '200':
          description: Список кофеен
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/CoffeeShop'
        '404':
          $ref: '#/components/responses/NotFound'
  /companies/{companyId}/menu/{drinkId}:
    delete:
      tags:
        - Companies
      summary: Удалить позицию меню
      description: Удаляет позицию из меню конкретной компании
      operationId: DeleteMenuItem
      security:
        - bearerAuth: []
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174001"
        - name: drinkId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Позиция меню успешно удалена
        '204':
          description: Позиция меню удалена
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  #-------------------------------------MENU-------------------------------------

  # /menu:
  #   delete:
  #     tags:
  #       - Menu
  #     summary: Удалить позицию меню
  #     description: Удаляет позицию из меню сети кофеен
  #     operationId: DeleteMenuItem
  #     security:
  #       - bearerAuth: []
  #     parameters:
  #     - name: drinkId
  #       in: query
  #       required: true
  #       description: ID напитка
  #       schema:
  #         type: string
  #         format: uuid
  #         example: "123e4567-e89b-12d3-a456-426614174000"
  #     - name: companyId
  #       in: query
  #       required: true
  #       description: ID компании
  #       schema:
  #         type: string
  #         format: uuid
  #         example: "123e4567-e89b-12d3-a456-426614174001"
  #     responses:
  #       '200':
  #         description: позиция меню успешно удалена
  #       '404':
  #         $ref: '#/components/responses/NotFound'
  #       '500':
  #         $ref: '#/components/responses/InternalServerError'
    
  #   post:
  #     tags:
  #       - Menu
  #     summary: Добавить позицию в меню
  #     description: Добавляет новую позицию в меню сети кофеен
  #     operationId: AddMenuItem
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: '#/components/schemas/Menu'
  #     responses:
  #       '200':
  #         description: Запись успешно добавлена в меню
  #         content:
  #           application/json:
  #             schema:
  #               $ref: '#/components/schemas/CreateResponse'
  #       '400':
  #         $ref: '#/components/responses/BadRequest'
  #       '409':
  #         $ref: '#/components/responses/Conflict'